DEP: rm method `Partition.preserves`
It appears that this method was never used, and its implementation
uses attributes that do not exist in the class, likely to have been
considered during initial design, but not implemented.
It was an attempt to abstract the notions in the method
`tulip.abstract.prop2partition.PropPreservingPartition.is_predicate_preserving`.

This method was introduced in [1].

[1] https://github.com/tulip-control/tulip-control/commit/a6fe42be1e4d1407cb7aa1cf1361fd3582e5f81b